+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `limberframework.database.connections.ServerConnection` class to connect to databases located on a server.
 - mysql+mysqldb driver for connecting to a MySQL database.
 - `limberframework.database.connections.ElasticsearchConnection` class to connect to Elasticsearch databases.
+- `limberframework.logging.log_service_provider` that provides the log service which returns an instance of `loguru._logger.Logger`.
 
 ### Changed
 - pre-commit to only pass files to pydocstyle from the _limberframework_ folder.